requirements.txt
flask
flask-cors
langchain
langchain-community
langchain-openai
langchain-groq
openai
chromadb
python-dotenv
pypdf
python-docx
docx2txt
langchain_chroma
SQLAlchemy
psycopg2-binary
geocoder
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
atlassian-python-api
pytz